feat: allow sorting and filtering on hidden columns
In render_output, we stopped using `filter_keys`, in order to be able to
sort and filter on hidden columns as well. Instead, we use
`reorder_keys` because filter_keys used to reorder keys and we want the
columns to appears in the order given by the arguments. Last but not
least: in `add_list_resource_command` we actually reimplement
`filter_keys` in order to remove columns that should never be rendered
(node._api for instance). First because we don't want users to display
them, and second because PrettyTable take in account the rendering of
hidden columns to calculate the height of each cells, which would
produce very high columns in some cases.