refactor: fix node list cells size
Filtering columns that should not be displayed prevent internal state of
an object to be rendered even if not displayed. PrettyTable adjusts the
size of the cells according to hidden columns as well. But hidden
columns should not be completely removed since we allow to filter and
sort on hidden columns too.